Move networking components to own module
