+++ mutiny-core/src/networking/mod.rs
@@ -1,0 +1,4 @@
+pub mod socket;
+
+#[cfg(target_arch = "wasm32")]
+pub mod proxy;
